SPA fix: add id=page-content to Layout main area. Clean Int32→Int in all examples.

+++ docs/src/routes/examples/index.jl
@@ -4,13 +4,13 @@
         P(:class => "text-warm-500 dark:text-warm-400", "Interactive examples built with Therapy.jl"),
         
         H2(:class => "text-xl font-semibold text-warm-800 dark:text-warm-200", "Counter"),
-        InteractiveCounter(initial=Int32(0)),
-        Pre(:class => "bg-warm-900 dark:bg-warm-950 text-warm-200 p-4 rounded-lg border border-warm-800 mt-4 font-mono text-sm overflow-x-auto", Code(:class => "language-julia", """@island function InteractiveCounter(; initial::Int32 = Int32(0))
+        InteractiveCounter(initial=0),
+        Pre(:class => "bg-warm-900 dark:bg-warm-950 text-warm-200 p-4 rounded-lg border border-warm-800 mt-4 font-mono text-sm overflow-x-auto", Code(:class => "language-julia", """@island function InteractiveCounter(; initial::Int = 0)
     count, set_count = create_signal(initial)
     Div(:class => "flex items-center gap-4",
-        Button(:on_click => () -> set_count(count() - Int32(1)), "-"),
+        Button(:on_click => () -> set_count(count() - 1), "-"),
         Span(:class => "text-2xl font-mono", count),
-        Button(:on_click => () -> set_count(count() + Int32(1)), "+")
+        Button(:on_click => () -> set_count(count() + 1), "+")
     )
 end"""))
     )
